Extract the REPL server from anthology into a new `colloquy` module
The interactive REPL had outgrown Anthology (a compiler wrapper). The server —
`Repl`, `ReplMacro`, `ReplModuleCompiler`, the Java `ReplBridge` registry and the
`soundness` re-export — moves verbatim from `anthology/repl` to a new top-level
`colloquy` library (`colloquy.core`), with the package renamed `anthology` →
`colloquy`. It still depends on `anthology.scala` for the `Scalac` it drives.

The REPL test suites move to `colloquy.test` (17 tests, all green); anthology
keeps a placeholder test. build.mill drops the `anthology.repl` component, adds
the `colloquy` library (with a `client` component for the forthcoming front-end),
and registers it in `allLibraries` and `groupCheck.excluded`.
