+++ build.mill
@@ -231,7 +231,8 @@ object groupCheck extends Module:
     "anticipation.http", "anticipation.log", "anticipation.opaque",
     "anticipation.path", "anticipation.print", "anticipation.text",
     "anticipation.time", "anticipation.url",
-    "beneficence.core", "beneficence.plugin", "dendrology.demo",
+    "beneficence.core", "beneficence.plugin", "colloquy.client", "colloquy.core",
+    "dendrology.demo",
     "ethereal.example", "exegesis.demo", "exoskeleton.args", "exoskeleton.core",
     "exoskeleton.rig", "frontier.plugin", "probably.coverage",
     "quantitative.units",
@@ -356,11 +357,7 @@ object anthology extends Library:
   object bundle extends Component(core, zeppelin.core, revolution.core, eucalyptus.core):
     override def scalacOptions = Task:
       super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,scala,proscenium")
-  object repl extends Component(`scala`, inimitable.core, coaxial.core, eucalyptus.core, stratiform.core):
-    override def scalacOptions = Task:
-      super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,scala,proscenium")
-    def mvnDeps = Seq(mvn"org.scala-lang:scala3-compiler_3:${scalaVersion()}")
-  object test extends Tests(`scala`, `java`, core, repl, spectacular.core)
+  object test extends Tests(`scala`, `java`, core, spectacular.core)
 
 object anticipation extends Library:
   object generic extends Component(prepositional.core):
@@ -539,6 +536,21 @@ object coaxial extends Library:
     override def scalacOptions = Task:
       super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,scala,proscenium")
   object test extends Tests(core)
+
+object colloquy extends Library:
+  object core extends Component(anthology.`scala`, inimitable.core, coaxial.core, eucalyptus.core,
+                                stratiform.core):
+    override def scalacOptions = Task:
+      super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,scala,proscenium")
+    def mvnDeps = Seq(mvn"org.scala-lang:scala3-compiler_3:${scalaVersion()}")
+
+  object client extends Component(ethereal.core, exoskeleton.completions, profanity.core,
+                                  coaxial.core, urticose.core):
+    override def scalacOptions = Task:
+      super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,scala,proscenium")
+    override def mainClass: Task.Simple[Option[String]] = Some("colloquy.repl")
+
+  object test extends Tests(core, spectacular.core)
 
 object contextual extends Library:
   object core extends Component(rudiments.core):
@@ -1483,7 +1495,7 @@ val allLibraries: Seq[Library] = Seq(
   abacist, acyclicity, adversaria, ambience, anamnesis, anthology, anticipation,
   apoplexy, austronesian, aviation, baroque, beneficence, bitumen, breviloquence, burdock,
   cacophony, caduceus, caesura, camouflage, capricious, cardinality, cataclysm,
-  cellulose, charisma, chiaroscuro, coaxial, contextual, contingency, cordillera,
+  cellulose, charisma, chiaroscuro, coaxial, colloquy, contextual, contingency, cordillera,
   cosmopolite,
   decorum, dendrology, denominative, digression, dissonance, distillate, diuretic,
   embarcadero, enigmatic, escapade, escritoire, ethereal, eucalyptus, exegesis,
